// src/api/model/dashboardInvoiceItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */
import type { DashboardInvoiceItemDtoStatus } from './dashboardInvoiceItemDtoStatus';

export interface DashboardInvoiceItemDto {
  id: string;
  number: string;
  companyName: string;
  date: string;
  amount: number;
  currency: string;
  status: DashboardInvoiceItemDtoStatus;
}